Find/fix array bounds bug in PoissonDisk
There's a few small changes to PoissonDisk, such as new constructors
that allow specifying the cutoff point to stop looking for more points
(specifying this is recommended in most cases where you don't want to
fill a space, since it helps performance drastically), and an
optimization that also makes certain that the pointLimit is respected.
But the big thing here is that the grid PoissonDisk uses is now
correctly calculated using the same technique it looks up items in the
grid later (both round now, instead of one rounding and the other
truncating). This prevents a nasty, though uncommon,
IndexOutOfBoundsException that could happen through normal usage. To
give an idea, it took over 800 attempts in a loop to find the bug
happening once, and I'm fairly certain it's fixed now because I ran the
loop one million times with no exceptions thrown.