Avoid briefly allocating huge amounts of RAM
Before, if pointLimit was negative then it would treat the limit as
0x7FFFFFFF, and would always make an ArrayList with a capacity of 1/4
the point limit. this is more than 512MB of empty array, and likely
closer to 1GB, which is past the heap limit on most JVMs even though it
was probably never used. Now the ArrayList always is initialized with
256 capacity, which may be a bit higher than is normally needed but not
massively so.