CrossHash update. NumberTools.
CrossHash's new algorithm passes SMHasher's difficult tests while the old one definitely didn't. NumberTools is used by CrossHash to convert double values to long bits, which NumberUtils doesn't implement in libGDX. It also has a few other nice and useful functions, and is super-sourced on GWT so it will act the same on desktop/Android/iOS and the web.